PropertyInsurance/requirements.txt
adium-theme-ubuntu==0.3.4
Babel==1.3
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.4.1
certifi==2017.4.17
chardet==3.0.3
click==6.7
cryptography==1.2.3
cycler==0.10.0
decorator==4.0.6
Django==1.11.2
djangorestframework==3.7.7
docutils==0.12
enum34==1.1.2
et-xmlfile==1.0.1
feedparser==5.1.3
funcsigs==0.4
functools32==3.2.3.post2
future==0.16.0
gdata==2.0.18
geocoder==1.33.0
get==0.0.39
gevent==1.1.0
googlemaps==2.5.1
greenlet==0.4.9
html5lib==0.999
httpie==0.9.9
idna==2.5
image==1.5.5
ipaddress==1.0.16
ipython==5.4.1
ipython-genutils==0.2.0
jdcal==1.3
Jinja2==2.9.6
lxml==3.5.0
Mako==1.0.3
MarkupSafe==0.23
matplotlib==2.0.2
meld==3.14.2
mock==1.3.0
numpy==1.13.0
olefile==0.44
openpyxl==2.4.8
pandas==0.20.2
passlib==1.7.1
pathlib2==2.3.0
pbr==1.8.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
post==0.0.26
prompt-toolkit==1.0.14
psutil==3.4.2
psycopg2==2.6.1
ptyprocess==0.5.2
public==0.0.65
pyasn1==0.1.9
PyChart==1.39
pycrypto==2.6.1
pydot==1.0.29
Pygments==2.2.0
pygobject==3.20.0
pyinotify==0.9.6
pyOpenSSL==0.15.1
pyparsing==2.0.3
pyPdf==1.13
python-dateutil==2.6.0
python-ldap==2.4.22
python-openid==2.2.5
python-stdnum==1.2
pytz==2014.10
PyWebDAV==0.9.8
PyYAML==3.11
query-string==0.0.28
ratelim==0.1.6
reportlab==3.4.0
request==0.0.26
requests==2.17.3
roman==2.0.0
scandir==1.5
scikit-learn==0.18.2
scipy==0.19.1
simplegeneric==0.8.1
simplejson==3.8.1
singledispatch==3.4.0.3
six==1.10.0
sklearn==0.0
sqlparse==0.1.18
subprocess32==3.2.7
tornado==4.5.1
traitlets==4.3.2
unity-lens-photos==1.0
urllib3==1.21.1
uTidylib==0.2
vatnumber==1.2
vobject==0.8.1rc0
wcwidth==0.1.7
Werkzeug==0.10.4
wkhtmltopdf==0.2
xlrd==1.0.0
XlsxWriter==1.0.2
xlwt==1.2.0
